AddedTopics

+++ app/src/main/java/com/example/cocktailmachine/data/Ingredient.java
@@ -5,13 +5,15 @@ import org.json.JSONObject;
 import java.util.List;
 
 public interface Ingredient {
+    //Reminder: Only liquids!!!
+
     public long getID();
     public String getName();
     public List<String> getImageUrls();
     public boolean isAlcoholic();
     public boolean isAvailable();
     public void addImageUrl(String url);
-    public float getFluidInMilliliter();
+    public int getFluidInMilliliter();
     public Pump getPump();
     public int getColor();
     public default JSONObject asMessage(){
